+++ server/db/ServerGroupDBInterface.go
@@ -9,7 +9,7 @@ type ServerGroupDBInterface interface {
 	GetServerGroupByName(name string) *model.ServerGroup
 	PostServerGroup(s *model.ServerGroup) (*model.ServerGroup, bool, error)
 	GetServerGroup(id string) *model.ServerGroup
-	GetServerGroupCollection(start int, count int, filter string) (*model.ServerGroupCollection, error)
+	GetServerGroupCollection(start int64, count int64, filter string) (*model.ServerGroupCollection, error)
 	DeleteServerGroup(id string) (*model.ServerGroup, error)
 	DeleteServerGroupCollection() error
 }
