+++ server/db/ServerServerGroupInterface.go
@@ -8,7 +8,7 @@ import (
 type ServerServerGroupInterface interface {
 	PostServerServerGroup(sg *model.ServerServerGroup) (*model.ServerServerGroup, bool, error)
 	GetServerServerGroup(id string) *model.ServerServerGroup
-	GetServerServerGroupCollection(start int, count int, filter string) (*model.ServerServerGroupCollection, error)
+	GetServerServerGroupCollection(start int64, count int64, filter string) (*model.ServerServerGroupCollection, error)
 	DeleteServerServerGroup(id string) (*model.ServerServerGroup, error)
 	DeleteServerServerGroupCollection() error
 }
