+++ server/object/dto/GetServerCollectionResponse.go
@@ -15,9 +15,9 @@ type ServerMember struct {
 
 // GetServerCollectionResponse is DTO.
 type GetServerCollectionResponse struct {
-	Start       int            `json:"Start"`
-	Count       int            `json:"Count"`
-	Total       int            `json:"Total"`
+	Start       int64          `json:"Start"`
+	Count       int64          `json:"Count"`
+	Total       int64          `json:"Total"`
 	Members     []ServerMember `json:"Members"`
 	NextPageURI *string        `json:"NextPageURI,omitempty"`
 	PrevPageURI *string        `json:"PrevPageURI,omitempty"`
