+++ server/object/model/ServerCollection.go
@@ -10,9 +10,9 @@ type ServerMember struct {
 
 // ServerCollection server collection.
 type ServerCollection struct {
-	Start       int
-	Count       int
-	Total       int
+	Start       int64
+	Count       int64
+	Total       int64
 	Members     []ServerMember
 	NextPageURI string
 	PrevPageURI string
