+++ task/object/dto/GetTaskCollectionResponse.go
@@ -24,9 +24,9 @@ type TaskMember struct {
 
 // GetTaskCollectionResponse Get task collection response DTO.
 type GetTaskCollectionResponse struct {
-	Start       int          `json:"Start"`
-	Count       int          `json:"Count"`
-	Total       int          `json:"Total"`
+	Start       int64        `json:"Start"`
+	Count       int64          `json:"Count"`
+	Total       int64          `json:"Total"`
 	Members     []TaskMember `json:"Members"`
 	NextPageURI *string      `json:"NextPageURI,omitempty"`
 	PrevPageURI *string      `json:"PrevPageURI,omitempty"`
